requirements.txt
librosa==0.11.0
matplotlib==3.10.7
moviepy==2.2.1
mtcnn==1.0.0
numpy==2.3.5
opencv-python==4.10.0.84
pandas==2.3.3
Pillow==12.0.0
scikit-learn==1.7.2
seaborn==0.13.2
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
